refactor: remove queries file
